backend/requirements.txt
# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PyTorch & Vision
torch==2.1.2
torchvision==0.16.2
timm==0.9.12

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# API & Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.3

# CORS & Middleware
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development (optional)
black==24.1.1
flake8==7.0.0
mypy==1.8.0

pymysql==1.1.0
cryptography==41.0.7